Cleanup of 2021 Day 8
Jesus, I'm not so pleased with my time anymore. I realized that I made a
critical mistake about the assumptions for this puzzle. The description
says that each line has every digit and only one of each digit. Knowing
this we can reason out which digit is which, by applying masks on the
remaining digits, until we have them all. It's not actually required to
map the segments at all, you just need to identify the digits.

What I'm more pissed of with though, is that I didn't just brute force
this. Looking at some of the solutions it's so clear that this would
have been bruteforceable! This is a note for next time then, always try
to bruteforce first, when under time pressure.

Another interesting approach I saw was to use frequency analysis. We
know that certain segments appear more than others, and so we can use
this to single them out.