2021 Day 11
Well this day was very boring, the task was way to easy. I got to use
my prebuilt "adj" function, which is nice, but other than that it was
shite.

The reason I was so slow to complete today, was that I was sleeping out
in the forest and it was really cold in the morning, so I didn't fancy
getting up and doing it. I did however have my pc with me and I did code
the solution with freezing hands and a computer that was visibly slow
from the cold.

I really hope that these tasks pick up at some point, because this is
getting very boring.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 11   04:22:30  13127      0   04:24:38  12822      0