2021 Day 12
Well, this is disappointing. I got an ok time for part one, but really
lost all my time to two factors on part two:

The first was that I thought we could visit every cave twice, but
apparently the task was that we could only visit ONE of the caves twice.
This was a mistake on my part, but not as bad as I thought it would be.
The largest example caught it and I was able to debug it.

The second mistake, which cost me much more time, was that I had a
fucking silly bug in my new system. I was not inheriting the boolean
which marks if we have seen a path twice already and this cost me a lot
of fucking time. #"#!"¤!)!?=!

It enabled one of the people at work to beat me for part two, which I
never would have expected. A good lesson in that you need to keep your
ego in check and to be careful of bugs in systems like this!

The day itself was rather mundane, but I guess it's better than nothing.
I really think this has been the worst AOC to date, by a long shot.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 12   00:08:41    239      0   00:28:05   1048      0