2018 Day 14 Part 2
Well, this turned out to be easier than expected? I sat down now to
solve this task, given that I thought you couldn't bruteforce it. I
implemented a dumb bruteforce solution, to have a baseline to compare
against and lo and behold the answer popped out, thought a bit slowly.

I think the mistake I made before was that I was only looking back by
the length of the input, but this doesn't work if you are incrementing
the string by two. Once this was fixed, it still takes some time, but it
does finish. I'm going to look at the megathread now to see how other
people solved it.