Small cleanup of 2018 Day 15
Ok, so the one off errors where caused by the fact that we are only
supposed to count complete rounds, and combat ends as soon as a unit
don't have any targets. This means that for most inputs, the last round
is going to be stopped midway through.

Once I fixed this, it now returns the correct answer for all inputs.
This is irksome, but at least it wasn't a big bug.

I'll go back later and clean this up, there is really just a bunch of
debug logging that needs to be removed. I'm not sure I'll touch the
logic all that much.