2021 Day 14
Second good problem of the year!

This is very similar to a lot of previous years, but there you have more
involved transformations. I spent most of my part 1 on actually getting
the most used character down. I need to get a grip on the "Counter"
class in "collections", I spent so much time on that shit :(

For the second half I quickly recognized that it was not going to work
to just bruteforce it. I quickly realized that you could just consider
each pair of characters alone, and simply evolve all pairs at once. I
then spent the next 20 minutes fiddling with two things:

Keeping track of the character count. In the beginning I thought I could
just add the first letter of each pair and get the answer that way. This
turns out to actually work, but due to another bug it did not work. So I
tried various things, but in the end I made a separate structure for
tracking the character counts.

This still didn't work on my example. This was very weird, until I
finally gave up and plotted the two solutions, bruteforce and my smart
solution, against each other and realized that my input has pair
combinations more than once. I had hardcoded each pair to 1 in my
initial setup, so that was why it was failing.

I now realize that my idea of adding the first letter in each pair would
also have worked, plus the end character, as it's the only one not seen
this way, but due to the other bug it didn't.

I wouldn't have placed either way, but a silly bug! I'm very happy with
having a nice day anyway! I will sit down and finally learn the damn
"Counter" class, as I fumbled so much with it on part one.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 14   00:06:51    287      0   00:27:04    778      0