2021 Day 15
ANOTHER LEADERBOARD PLACING!

The morning started of on a sad note, as a winter half marathon I had
been preparing for got canceled due to the new omicron variant. I won't
get all the money refunded, so that was a bit of a bummer.

The day then got a bit brighter, when I placed 57 on part two of todays
puzzle! It was very unexpected, given that I placed 185 on part one. I
guess there was a couple of factors working in my favor here:

The first is that I had already written a procedure for picking
neighbors of a square. This sped up writing part one by quite a bit, but
should not have any affect on part two.

On part two I was a bit slow to decide, but I decided not to actually
expand the grid, but rather just calculate the value. I also noticed
quite quickly that we wrap around to 1 and not 0. This was hard to do
mathematically, so I decided to just do it with a range expression.

I will go back and switch over the range expression to something
mathematical, I'm sure there is a solution for it.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 15   00:06:59    185      0   00:12:27     57     44