Cleanup of 2021 Day 15
So, it turns out you don't need to do the finicky mod wrap around at
all, it's enough to just shift every value by -1 and then just add 1
when you are adding up the risk values. This makes it possible to simply
use normal modular arithmetic with 9 as the modulus.

Part two is still a bit slow for my taste, will have to investigate that
more.