2021 Day 16
A good parsing problem, makes for a good day!

This was a step up in difficulty from the previous days, but I spent
most of my time trying to understand the problem description.

I spent much more time on part one due to a small bug in my code. I
didn't get the fact that the zero padding that is talked about for
literal values, have happened already. So we do not need to concern
ourselves with it. I thought we had to pad the number until it aligned
or some stuff, which made this a bit harder.

Once I fixed that, everything else really fell into place. My code is
ugly as hell, because it was rushed. For the first task, I didn't bother
to do the return values properly, I just printed out the version headers
and then added them together on the commandline.

Part two I quickly solved, though I'm sure it could have been done
quicker.

Overall a very good day, looking forward to cleaning this up later.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 16   00:31:20    399      0   00:38:02    362      0