2021 Day 17
Well, this was an ok fun day! I did ok on the scoreboard, I'm happy with
my performance in that area.

I basically bruteforced this, and guessed at the input ranges.
Everything basically worked on the first try and the only bug I had was
that in part two, I was testing for if the max_y was true when evaluated
as a boolean. This misses cases where it is 0, but I quickly got that
fixed.

I also didn't waste anytime parsing the input format, I just hardcoded
it.

I think the real fun in this one is going to be to optimize this, and/or
solve it analytically. I will try to do this by myself first and then
later read the answers on the scoreboards.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 17   00:11:45    294      0   00:13:29    133      0