Cleanup of 2021 Day 17
So this didn't clean up very well, but I managed to make some smart
choices about how to cut of possibilities.

This day was ok fun, but I am a disappointed, because I actually think
there is a way to prove the answers to either part 1 or part 2. There
will always be an initial dx which will line up perfectly with the
target area and no matter how high we throw it, it will fall straight
down onto it.
For these initial "dxs", it's just a matter of tuning the initial dy
such that it also overlaps the target area.

This comes down to solving:

y_m(t) = - (t * (t - 1 - 2*dy)) // 2

for a range:

a <= y_m(t) <= b && 0 <= t

where a, b and t are integers.

I don't think there is any way to get a bounded solution here,
so we cannot know if there are a bounded number of solutions. This
breaks both part one and part two of today's puzzle.