Proper cleanup of 2021 Day 17
Ok, so I was very wrong in my previous statement here. You can bound the
upper limit on the velocity, as the ball is always going to come down to
y=0 again, with (-dy)-1 velocity, which means that we can cap on that.
Once that is done, part one becomes a closed form oneliner and part two
goes very quickly.

Some of my tricks still help on part two, so I'm glad I put in the time.