2021 Day 18
What a nice day! Very bespoke challenge and not trivial to solve.

The real hard task was to figure out the "next" number in each
direction. In the end I was able to do it first try, without any bugs!

I'm surprised by my ranking here, I was expecting it to be lower, as I
spent a lot of time figuring this out. But it appears everyone else did
so too!

One really nice thing here is that he formatted the inputs so they could
simply be "evaled", which is really nice for me! I'm thinking I'll use
something else for the final solution, but it really was rather nice!

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 18   00:50:48    277      0   00:53:34    251      0