Cleanup of 2021 Day 18
I looked at the solution thread and I'm surprised to see that many
people solved this using strings! I've cleaned up my solution, basing it
on one posted by "1vader". The main difference here is that instead of
mutating the input, we are returning the changed tree. This makes it
easier to reason about side effects and it's quite a bit shorter, but
it's not faster.

I'd be interested in seeing how you can make part 2 of this fast.