2021 Day 19
Geez, Louise! What a day! This is looking more like the AoC that I know!

I don't really care about scoreboard here, since I'm so far off it, but
the task itself was interesting. The method I use for figuring this out:

Use scanner 0 as the "true" orientation, and it's position is at
(0,0,0). Then we loop through the other scanners and find ones that
share 12 probes with it. We determine this by looking at the distance
between pair of probes. We assume that if two probes have 11 equal
distances to other probes, then that probe and those 11 probes must be
equal. This isn't necessarily true, but for AoC it is usually safe to
assume such things.

Now that we have found a potential match, we attempt to find the
orientation and offset of the second scanner, relative to (0,0,0). We do
this by finding a pair of matching probes in each set. We find these by
comparing their distance to each other. Now that we have that, we figure
out the rotation. I do this by just matching up the three axis on the
distances. One can also use proper rotation and try all possibilities.

Once we have the rotation and the offset of the second scanner, we also
know it's location, which is just the offset. We then "correct" the
seconds scanners sensor readings, by first rotating them and then adding
the offset. Once this is done, we add it to the list of probes we want
to check next.

We continue doing this until there are no more probes left to examine.

I managed to get the matching metric, 11 equal distances, very quickly,
but struggled on the offset and orientation part. I had errors all the
way, and I tried to use proper rotation and so on to fix this. In the
end I think I must have had some sort of logic bug somewhere, because my
rotation code that I wrote initially is correct.

I initially tried to do the matching first in one pass and then the
correction in another. I didn't get this to work and things where
getting to muddy. I decided to start anew, from the start, so I started
on the "solve2" method. I didn't really change anything here, besides a
sign in the "orient" function and not doing the matching pass first.
This somehow solved it and I think I spent 10 minutes from starting
"solve2", to getting both stars.

This shows that I should have done this earlier, but I could not have
known this. It will be interesting to go back and optimize and clean
this up, but now I got to do some chores :)

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 19   02:47:04   1243      0   02:49:42   1108      0