2021 Day 20
God damn it, I was so close to being on the leaderboard for part two
today! And I would have been on the overall leaderboard if it hadn't
been for a simple "continue" statement I put in to try to optimize a
bit. Jesus!

The insight here is that in our inputs, the "image enhancement
algorithm" has a "#" in the 0th slot, and a "." in the 511th slot.
This means that it will alternate between every pixel being off, except
for those that are set to on specifically, to every pixel being on,
except those that are set to off specifically. This means that the
question being asked only makes sense on even numbers.

I realized this quickly and wrote an implementation for this, using a
"defaultdict" which alternated between defaulting to true or false,
depending on which iteration I was on. I then only considered the
especially set pixels from the last iteration.

The mistake I made was that when I was writing the "to_consider" list, I
didn't include spots which was false, but this leads to wrong behavior
when the default is that all pixels are on.

It runs ok fast, and part two was just to adjust the number of
iterations. I will try to implement this smarter now.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 20   00:18:35    153      0   00:19:01    102      0