2018 Day 16
Fun day! I really like these sort of reduction problems, so this played
right into my hand. I would have placed on the leaderboard for both
parts, if I had performed like this during the actual year!

The code is pretty self explanatory, the only weakness right now is that
it will error if we try a register out of bounds. An exception catch
would help with this.

Part one: 00:19:03
Part two: 00:08:20