2021 Day 22
Well, what a day! This is the most challenging task this year I think, I
have not done much of 3D programming like this. It is something that I
want to get better at however, so I really should look into this!

I solved part one fast enough, but I already knew what part two was
going to be. I initially tried to find a library to do the collision
detection and this is where I spent most of my time.

None of them did really what I wanted and it was too much new
information to look into to get a good score. After about 2 hours 45
minutes, I started trying to hack together something with the excellent
"intervaltree" library.

I use 3 levels of the intervaltrees and use that. It took a little while
to figure out how to do everything, but in the end it worked, to my
surprise. I think I should have done this right away, as I briefly
considered it, but judged that it would not work and be hard.

I think this is the best takeaway here, that you should try to stick
with what you know until that doesn't work, then look for new libraries.
Had I started trying to hack something together, I would have been able
to solve the task and maybe place.

I also had a OOM error in the middle of this, trying to apply my part
one solution to part two. This didn't work out to well :P

I'm really excited to see how others solved this, and to clean up my
solution files a bit!

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 22   00:08:03    283      0   03:13:50   2014      0