2021 Day 23
It's my birthday and for a present, I got a really nice task!

I realized quite quickly that this is really some sort of "towers of
Hanoi" variant, that can be solved without doing any sort of graph
search. I also realized that trying to be smart about this, would lead
me to spending much more time than if I just go with being stupid about
it :P

I ended up doing a BFS search of the board for each creature each turn
first, then I filter out the moves that are not valid for a creature to
make and the remaining possibilities is what I go off. I use an A*
search to move through the possibilities with a custom made, simple
heuristic.

I hardcoded the animal names and number, which would be a mistake in
part two, but whatever. I'm quite pleased with my part one, I did it as
quickly as I think I could.

Part two was a real nice challenge, but I made quite an important
mistake, which is why I spent almost an hour on part two, instead of 10
minutes. I forgot to update the "done" check, so the algorithm didn't
stop.

I would have found a solution much earlier if not for this, but this did
lead me to try a bunch of different approaches and small optimizations.
One of these is to convert from a pure "BFS" to a A* for the node
exploration. I also added in the concept of "moves" and such. I also
realized that it doesn't really matter which instances of the different
type of creatures are where, so I reduced the "seen" memory usage quite
a lot there, by ignoring the order as I went in.

The reason I realized that I had wrong "done" function is actually that
I had put process to run on my input in the background and it completed
without finding a solution. Once I saw that, it took 2 minutes to find
the answer.

I'm very happy with today, my placing could have been better, but I
don't think I would have made it to the leaderboard anyway, so I'm ok
with this. The task was fun and I had a blast!

For part two, I didn't change my board parsing, I just manually created
the extended board. This means that as of now, part two has to be run
against the "inp.txt" file, but I will change this when I clean up!

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 23   00:46:37    703      0   01:39:37    352      0