Cleanup of 2021 Day 22 Part Two
So it turns out that you can be much smarter about this than I was.
There is really no need to keep track of splitting the cubes, you can
simply just calculate the overlap with existing cubes, and tally that
up.

This runs about as fast and is soooo much simpler to code! I will update
part one to use this logic too!