2021 Day 24
What a task it was today! The toughest task of the year and maybe one of
the harder I've solved in AoC overall!

I spent a long time on this one and I tried multiple different avenues
when it comes to solving it.

The first approach I took was to try solve it by parsing the
instructions into a tree and then solve on that, creating an AST for
much the same purpose. I tried using z3 and sympy without making any
headway.

After maybe 3 and a half hours, I gave up on a general approach and
decided to try to reverse engineer my input directly. I translated it
into a python program using some sed commands and started going through
each step.

Here I realized that there is really is the same instructions 14 times,
but with 3 different variables, let's call them a, b and c. The output
from each round is the "z" register and the round is only dependent on
z, a, b, c and the input for this round, w.

Once I was able to isolate this, I then started reducing the round
function further, until it was something I could reason more
analytically about. I wasn't able to make some elegant solution as to
why this is the case, but in the end it comes out so that the only way
to reduce the "z" register in a round, is if the input number and the
"b" variable fullfills the equation:

((z % 26) + b) == w

where "w" is the input variable. I then guessed that we would take every
opportunity we could to reduce the variable and I wrote a function
around this. It feels like there should be some proof for this, but I'm
not able to work that out now.

I also don't know if all inputs follow the general pattern that I had,
and that you can solve this task by reading the abc's straight from the
input at specific places. I'll look more into this later.

One guy at work beat me today, with a time of 2 hours 58 minutes. I am a
bit mad about that, a blow to the ego :P

A great day, mixing reverse engineering and analytical thinking! I love
it!

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 24   04:58:40   1901      0   04:59:14   1793      0