2021 Day 25
Well, a somewhat anticlimactic ending to an anticlimactic year. The task
today was really simple, I could have placed much higher, but I made one
silly mistake, which was that I did an "x = new_x" mistake, where the
mistake was: "new_x = new_x". *sigh*

This year has been a year of extremes. Some of the days where some of
the best of all time on AoC, but most of them were just trivial. I guess
this was an attempt to onboard more people and make the processes
smoother. Everyone deserves to feel the joy of problem solving, so if
this is the price to pay to get people hooked, then maybe it's worth it.

I've placed multiple times this year, 4 to be exact, and I've
accumulated 203 points. I would never have dreamed of doing so well, but
I will not stop practicing, so that I might place higher next year!

There are still some tasks that I should go back and optimize, but other
than that, I think that it is for this years python solutions. I still
have to go through and solve them all effectively with C++, and try to
use SIMD.

I'm also thinking that I will bite the sour apple and go back to rust
again too. It might just be the language of the future and so I should
be using it more frequently.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 25   00:12:54    303      0   00:12:58    261      0