2022 Day 1
It's Advent of Code again and I'm excited!

There are some dampeners this year though. I'm moving in the beginning
of the month and I'm going to some personal stuff, so I haven't been
able to practice at all for this year. I'll still try to get up early
and get after it, but I've decided to be more relaxed this year, and not
hold myself to the same standard I did last year.

Enough of that, on to the task. I think it was OK, though a bit
uninspired. I spent to long coding this, but I always do. It will be a
nice little task to clean up though.

Score:
      -------Part 1--------   -------Part 2--------
Day       Time  Rank  Score       Time  Rank  Score
  1   00:01:47   301      0   00:03:46   663      0