2022 Day 2
Coding on the laptop today and kinda felt it. I think I solved the task
quite elegantly, but I was slow in typing and the approach required
mentally adding up the possibilities.

The score on the second part was quite suprising, but that's always nice
:)

Score:
      -------Part 1--------   -------Part 2--------
Day       Time  Rank  Score       Time  Rank  Score
  2   00:05:49   667      0   00:08:18   377      0