2022 Day 3
Daaammnnnnn it was hard to wake up today.

This was mostly a challenge in parsing the problem text fast enough, and
also looking up the and operator in python sets and "islower".

I'm going to look into more-itertools again, because it has the grouper
method I think, which would have made this real easy.

Score:
      -------Part 1--------   -------Part 2--------
Day       Time  Rank  Score       Time  Rank  Score
  3   00:05:17   655      0   00:08:25   498      0