2022 Day 4
Another kind of uninspired day, but I did OK. I really should invest in
creating a routine for just grabbing all the numbers from a line.

I got the first one easy, but had to google for the second one, but I
knew there was an easy way. I found a couple of ways to do this,
but here is a very good explanation:

https://stackoverflow.com/a/325964
https://stackoverflow.com/a/3269471

It won't really be hard to clean this up, but we will do it as always.

Score:
      -------Part 1--------   -------Part 2--------
Day       Time  Rank  Score       Time  Rank  Score
  4   00:03:28   530      0   00:05:40   640      0