2022 Day 5
Ok, so now the tasks are becoming a little better, at least the time
there was a fun parsing aspect to the problem, though the implementation
turns out to be rather trivial in python and with this size. I think
this marks itself as a good task to try to optimize, as what I'm
currently doing is really ineffective.

I spent most of the time here parsing the task, I think it might have
been faster to just hardcode in my stacks and then go from there, but I
was a bit stubborn. It was of good help though, as it allowed me to
detect the fact that I thought we could move multiple crates in part
one, but it's actually just one at a time.

The change from part 1 to part 2 was very minimal so the delta looks
nice. I hope this is a sign that there will be more interesting
problems.

Score:
      -------Part 1--------   -------Part 2--------
Day       Time  Rank  Score       Time  Rank  Score
  5   00:12:41   925      0   00:13:17   612      0