2022 Day 7
Well, finally a good task! This was very interesting!

It's a very interesting task, especially if you only see part 1. You
have two choices, which is either to just focus on the files, or to
actually build a tree structure with empty directories also. I chose to
go for just files in the end, but I spent a bit of time deciding this,
which turned out to be a mistake in the end.

What I spent the most time on here, was actually building the logical
tree. I should have been better here, I simply couldn't get my head
around it for a little while, but it's really quite simple. You can just
keep a list of the files and for each of them update the size of all
directories above. This is what I ended up doing, but I got side
tracked, thinking I should do this more effectively by building up the
total size from the bottom up.

This is more effective, but it's also harder. There are so few files in
this tree, so in the end I just went for a more bruteforce solution. I
might clean this up more later.

The second part came quite easy, once I had parsed the task correctly.

I didn't do to well on the leaderboard, mainly due to using so long  on
building the logical tree. I blame this on early in the morning, but
more importantly not having done the important work of warming up with
graph structures in python before AoC this year.

Either way, this was a really good day, hoping for more days like this!

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
  7   00:30:11   2041      0   00:34:46   1667      0