Improve scalability of 2022 Day 7
I switched to using a topological sort of the nodes, and only summing
together the nodes after building the tree. This would scale better for
larger and deeper trees, as it only processes each node once.

It is overkill for our inputs, but it was fun to do, especially since
python has builtin support for it, in the "graphlib" package.