2022 Day 8
Well, now we are getting somewhere! It's not as cool as yesterdays task,
but it's a step in the right direction.

I woke up late for this one too, I've got to go to bed earlier, but I'm
ok pleased with my time here. I figured out an approach quite quickly,
but I spent a bit to long parsing the task.

On part two I had a small bug, but I figured it out quickly enough.

I think there will be a bunch of cool optimizations for this one, and
also visualisations.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
  8   00:23:39   4041      0   00:37:11   3218      0