Cleanup of 2022 Day 8
Didn't really change much here, I looked at the subreddit and people
haven't really posed anything that is blood optimized yet.

I see a lot of them went for multiple directions as delta vectors,
instead of just hardcoding it like I did, and I think that cost some of
them a lot of time. A lesson for the future, dumb is often faster than
smart when there is not much to copy and paste.