2022 Day 9
I managed to wake up on time today, which is quite nice! I've tried very
hard to manage this each day, but I've messed it up the last 2.

No matter, let's talk about today's task. This was a fun task to solve,
because it was a bit unorthodox. I took quite a while to understand the
wording of the rules, but once that was done, I decided to hardcode
almost everything.

This actually worked on the first try, which was really nice. I'm sure
I'll find some really elegant stuff in the megathread today, as people
actually solved this much faster than I did.

I should have spent way less time on part two, but there is a sentence
in the puzzle saying: "be careful, new type of motions are possible",
which I worried about, so I modified some of my logic from part one. But
this is not necessary I think. First thing I'm doing once this is up, is
to revert those changes and see if I still get the right answer.

I'm really liking the ramp up this year, towards the more grid based
problems that AoC is known for!

A heads up for tomorrow, I will be out for a party today and won't have
access to a laptop when I wakeup. This means that my time for part 10
will be very bad, but that's just the way it is ;)

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
  9   00:16:01   1168      0   00:27:42   1087      0