Cleanup of 2022 Day 9
Ok, so I learned a lot from the megathread today, but I also figured out
some stuff on my own!

The first discovery was that my initial coding of the movement rules
where correct for both part one and two, without any modification. After
seeing that, I decided to simplify it, which turned out to be quite
elegant. We can just do the normal check if the head and tail are
touching, and if not, the delta to move the tail is simply the sign of
the delta between the head and the tail in each direction.

I simplified that up quite a bit, but wasn't able to find any real
shortcuts after that. So I opened up the megathread and I found two
things there, which was of interest.

The first was that in part one, you can apparently just move the tail to
the heads last position, when they are not touching. This makes sense if
you think about it, but I don't think I would have gotten it on part 1,
if I did this test a 100 times. So that was clever and it explains why
we got the warning in part 2!

The second was that it's easier to represent the grid positions as
complex numbers. I knew this from before, but had kinda forgotten it. I
converted my solution to use complex numbers, and the code became a lot
cleaner, mostly because of the builtin addition operator.

A nice little task, let us hope there is more like this coming :)