2022 Day 11
Well, this was a cool day, with a very nice twist on part two.

I slept through the alarm, so I woke up about 16 minutes past 6. I
wouldn't have placed the way I was doing stuff, but I am getting very
lax with this.

For part 1, I spent most of the time parsing. This could have been made
faster by imputing the data values manually, but I was tired, and didn't
trust myself. In general, this part could have gone faster.

The simulation itself did not take very long to write, but I had to take
some time to debug, because I was indexing the array of monkeys with the
wrong variable.

Part two is where it got interesting. I must admit here, that the moment
I saw "find another way to keep your worry levels manageable", I knew
that this was going to be some sort of modulus task. I immediately
thought of the Chinese Remainder Theorem, as I've worked with that
before. I just tried this out on a hunch, and it worked!

I think this would have taken a damn long time if I hadn't known about
it before hand though! I will sit down and write an explanation of this
in the task, so I'll know what was going on, when I come back.

The only other cool thing here, is that I ended up using "eval" for
evaluating the expressions. This is obviously slow, but it gets the job
done.

The real time taken:

Part one: 00:25:38
Part two: 00:06:20

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 11   00:47:03   4246      0   00:53:23   2183      0