2022 Day 12
Well, today was rooouuuggghhhh. I'm actually rather ashamed, graph
algorithms used to be my bread and butter, but I found myself not even
remembering basic shit today. I'm going to blame a lack of preparation
and early in the morning, but secretly maybe I've just gotten a bit lax
with my algo skills.

Well, let me tell you the tale of my incompetence. I begin with opening
up part one and quickly realizing that, yes, this is a graph problem. I
also have the idea that BFS will be good enough here, and I start
implementing it. The implementation is slower than I would like and I
make multiple mistakes. Some indexing ones, but the most major one is
that I don't account for the fact that we can go "down" as long as we
want, but we can only go one up. I read it as you can only go down 1
level as well.

I don't realize I'm making this mistake, so the example works, but it
times out on the actual input. I try out some different things, but I
can't get it faster, so I start doubting my assumptions about the task,
but also about the fact that BFS is equal to Dijkstra's algorithm for
unweighted graphs.

In the end, I chose to bite the bullet and download networkx, and start
reading the documentation. I fumble around there for some time, until I
realize that I have to use "DiGraph" to get a directed graph. Once I do
that, I get told that there is no path between the two. I spend more
time debugging this, but in the end go back to the task and really
reread it.

That's when I catch the hight thing. The worst part, is that I remember
now, that the first time I read it, I noticed the asymmetry and thought:
"Ok, gotta remember that and make the graph directional". I made the
graph directional, but I forgot about the checks for building the graph.

When I fixed that, I got the first part. Then for the second part I
again tried to use my bfs implementation, but for some reason that I
still don't know, it's to slow and didn't return the right answer? I'm
sure it's going to end up being a minor thing, but it's very
frustrating.

It's an interesting problem, multiple source, single destination,
shortest path. I haven't actually seen this before, but from what I see,
it would be an easy thing to modify BFS to solve, just add more nodes in
the Q in the start.

Ok, but I thought that one way to do this was to start from the end and
just terminate on first "a" seen. But this did not work, because now
that we are going the other way, I have to reverse the condition for
changing levels, but I didn't remember that before after.

I then just looked up BFS on networkx and eventually found "bfs_layers",
which does almost exactly what I need. After a bit of messing around
there, I got the answer.

This was quite humbling, I guess it really wasn't my day and the
implementation of these algorithm really is "use it or lose it".

While writing this, I realize that we can actually do better than BFS
here, because we have a heuristic that BFS does not have. We can always
just go 1 level up, and so we know the minimum amount of steps needed is
at least 26 minus the current level. I'll sit down now and try to
implement this.

I'm also very curious as to why my BFS implementation didn't work, but
I'll sort that out as well.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 12   00:32:54   2296      0   00:49:11   2976      0