Cleanup of 2022 Day 12
OK, so it seems I was just tired when I did this. The BFS idea works
right away when you have the right logic for finding the available
next nodes. The same goes for the bfs search from the end in part
two.

I still want to try out the A* algorithm to see if it performs better,
but I wanted to store this here, as an example of what I would have done
normally.