Optimize 2022 Day 12
This is a very small optimization, moving from BFS to A* for both parts.
I did a simple count of the number of loops we did in each, and the
result where:

p1: 5789 -> 5766
p2: 3664 -> 3628

Which is very small reductions and it might in fact be slower than the
previous versions for these inputs, given the overhead of using a heap
versus just a simple list.

This is however guaranteed to perform quite well on almost all inputs,
which is nice.