2022 Day 13
Well, today was quite fun. It is a rather straightforward task, but so
early in the morning I wouldn't put it past me to simply mess up. But I
didn't. I spent way to long understanding the sorting rules, but the
solution I came up with was rather nice.

And also, thank god for "eval". I know there are safer alternatives, but
in the heat of the moment, it's really nice being able to just go full
"eval".

I also learned about "functools.cmp_to_key". This makes part two rather
easy.

Overall a nice day, but it's more a matter of dotting all the i's and so
forth.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 13   00:15:17    494      0   00:21:06    495      0