2022 Day 15
FUUUUUUUUUUUUUUUUUCCCCCCCCCCCCCKKKKKKKKKKKKKKKKKKKKKKK

I decided that I would sleep in today, to try to get back on my feet
sleep wise. I thought to myself: "I'm not going to get on the
leaderboard anyway". I guess I should have known better than to jinks
myself like that, but nothing to do about it now.

My real times where:

Part one: 00:09:19 (Rank 58)
Part two: 00:09:19 (Rank 19)
Total:    00:18:39

I even had a stupid mistake on part one, where I didn't change the y
from 10 to 2000000, and so that cost me 60 seconds.

This one really hurts. But I guess I should just take it with me that I
would have placed on at least one task this year.

So talking about the task itself. When I opened it up, I was a bit
surprised to see that this was very similar to an earlier years task. I
cannot find it right now, but that one was quite hard. I quite quickly
figured out the ruleset and focused only on the row that we are
interested in. I just added all places that where closer to the senors
than the beacon on that row to the set.

I made one silly mistake, which as noted above, was to forget changing
the row from 10 to 2000000 which cost me 60 seconds. I realized it
almost immediately and so was able to fix it and just wait to be able to
enter the input.

For part two, I very quickly decided to go for using z3 to solve it,
seeing as it really was just a matter of solving a couple of equations.
I had to google how to use z3 in python again, and it took a little
time, but it worked almost straight away, after i had figured out how to
do abs. I really love z3 for these kinds of tasks, it's almost like a
cheat code to be honest.

I don't really know how I would have done this without it, but I guess
that will be a fun task to try to figure out.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 15   05:12:15  12374      0   05:21:36   7458      0